do not assume docker socket access

+++ actions-runner/docker-entrypoint.sh
@@ -5,8 +5,10 @@ set -eou pipefail
 chown -R runner /app
 
 # make sure the host GID gets remaped to this container's docker GID
-DOCKER_GID=$(ls -ln /var/run/docker.sock | awk '{print $4}')
-groupmod -g "${DOCKER_GID}" docker
+if [ -f /var/run/docker.sock ]; then
+  DOCKER_GID=$(ls -ln /var/run/docker.sock | awk '{print $4}')
+  groupmod -g "${DOCKER_GID}" docker
+fi
 
 # github actions runner setups a svc file after install
 # so if that doesn't exist,'
